factor functions out of fill_predictedscore_table into prediction_utils
